[Fix] 폴더링 수정 및 불필요한 파일 삭제
