+++ Winey/Winey/Source/Global/Supports/SceneDelegate.swift
@@ -15,7 +15,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         let window = UIWindow(windowScene: windowScene)
         self.window = window
-        let tabBarController = TabBarController()
+        let tabBarController = MyFeedViewController()
         window.rootViewController = tabBarController
         window.makeKeyAndVisible()
     }
